fix: Code review fix for useStyles
