+++ components/transactions/utils.ts
@@ -1,13 +1,6 @@
-import { ChainStateEnum } from "../state";
-import {CHAIN_OPTIONS} from '../components/transactions/constants';
-
-export const middleEllipsis = (text: string, start = 9, end = 9) => {
-  return `${text.slice(0, start)}...${text.slice(Number(`-${end}`))}`;
-};
-  
-export const endEllipsis = (text: string, total = 18) => {
-  return `${text.slice(0, total)}...`;
-};
+import { ChainStateEnum } from "@/state";
+import {CHAIN_OPTIONS} from './constants';
+import { replaceAll } from "@/utils";
 
 export const fetchMainAddress = (address: string) => {
   if (address.startsWith("ELF_") || address.endsWith(`_${ChainStateEnum.AELF}`) || address.endsWith(`_${ChainStateEnum.tDVW}`)) {
@@ -17,11 +10,6 @@ export const fetchMainAddress = (address: string) => {
   }
   return address;
 };
-
-export const replaceAll = (str: string, find: string, replace: string) => {
-  const regex = new RegExp(find, 'g');
-  return str.replace(regex, replace);
-}
 
 export const getChainFromAddress = (address: string) => {
   if (address.endsWith(`_${ChainStateEnum.AELF}`)) {
@@ -33,15 +21,13 @@ export const getChainFromAddress = (address: string) => {
   } 
   return "";
 }
-
+  
 export const getFormattedAddress = (address: string, chain: ChainStateEnum) => {
   if (!address.startsWith("ELF_")) {
     address = "ELF_".concat(address);
   }
-
   if (!address.endsWith(`_${ChainStateEnum.AELF}`) && !address.endsWith(`_${ChainStateEnum.tDVW}`) && !address.endsWith(`_${ChainStateEnum.tDVV}`)) {
     address = address.concat(`_${chain}`);
   }
   return address;
 };
-
